+++ src/sort/multiset.rs
@@ -684,11 +684,7 @@ impl PrimitiveCommon for SumMultisets {
 }
 
 impl PurePrim for SumMultisets {
-    fn apply<'a, 'db>(
-        &self,
-        mut state: crate::PureState<'a, 'db>,
-        args: &[Value],
-    ) -> Option<Value> {
+    fn apply<'a, 'db>(&self, state: crate::PureState<'a, 'db>, args: &[Value]) -> Option<Value> {
         let mut data = MultiSet::<Value>::new();
         let ms_of_ms = state
             .container_values()